Move Load in file
