+++ src/libs/antares/study/scenario-builder/rules.h
@@ -39,6 +39,7 @@
 #include "solarTSNumberData.h"
 #include "HydroTSNumberData.h"
 #include "WindTSNumberData.h"
+#include "LoadTSNumberData.h"
 #include <map>
 #include <memory>
 
@@ -61,7 +62,6 @@ public:
     //! Map ID
     using MappingID = std::map<int, Ptr>;
 
-public:
     //! \name Constructor & Destructor
     //@{
     /*!
@@ -109,7 +109,6 @@ public:
     // When current rule is the active one, sends warnings for disabled clusters.
     void sendWarningsForDisabledClusters();
 
-public:
     //! Load
     loadTSNumberData load;
     //! Solar
